🚚 Rename file to `token-counts` to align it with the unit test file
