+++ jest.config.js
@@ -1,5 +1,7 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
+const config = {
+  preset: "ts-jest",
+  testEnvironment: "node",
 };
+
+export default config;
