+++ src/index.ts
@@ -1,2 +1,2 @@
 export * from "./functions.js";
-export * from "./token_estimation.js";
+export * from "./token-counts.js";
